requirements.txt.txt
**requirements.txt**

tensorflow
streamlit
numpy
pillow
matplotlib
seaborn
scikit-learn
pyngrok